refactor(scanner): merge unexpected char and symbolic atoms
Previously, the case where we were checking for an invalid symbolic
atom was unreachable because of a check landing the code into
symbolic atom land. This was not robust because we needed to keep
in sync the check and the symbolic atoms switch. Now, instead,
the code lands directly in symbolic atoms checking.

As a plus, we now have 100% of coverage.