refactor(parser): always use parseWithFlags
Here the idea is that which is legal depends on context and we
are going to use flags to control this.

It was already kind of like that, but the distinction bewteen
statements and expressions was artificial and the new code, while
being lower level, is also more clear and explicit.