feat(scanner): recognise .. and ...
I still don't have a clear use case for .. but I have at least
one for `...`, which is that of leaving the body of a lambda
unspecified. In turn, we need this to write down type classes
for functionality implemented in the interpreter.

For example:

```scm
(typeclass Num (T)
  (define + (lambda (a b) ":: (Callable (T T) T)" ...)))
```

If we are able to write this, then the typechecker does not
need to hardcore the typeclasses implemented by the interpreter,
rather we can load them from an external file, which will
probably be named `./stdlib/runtime/runtime.brs`.