feat(typechecker): load most of the stdlib from disk
This chance allows us to update or tweak the standard library
definitions without recompiling the buresu command.

Likely there will be additional changes. The current way in
which we load the standard library is that only the typechecker
loads it. However, we will also need to load it from the
evaluator as well. However, because the evaluator does not
allow defining overloaded functions at the moment, attempting
to load the current stdlib will fail.