feat: implement the declare expression
Declare allows the typechecker to know about built-in functions
in a way that the evaluator can see the same expression and avoid
thinking it needs to define a new function with that name.

In principle, this allows the evaluator to also interpret the
standard library, since now it will just skip over the declare
expressions and keep its previous definitions in place.